+++ packages/imperative/src/config/__tests__/ConfigAutoStore.unit.test.ts
@@ -12,7 +12,7 @@
 jest.mock("../../logger/src/LoggerUtils");
 jest.mock("../../events/src/ImperativeEventEmitter");
 
-import { AbstractAuthHandler } from "../../imperative";
+import { AbstractAuthHandler } from "../../imperative/src/auth/handlers/AbstractAuthHandler";
 import { SessConstants } from "../../rest";
 import { ImperativeConfig } from "../../utilities";
 import { ConfigAutoStore } from "../src/ConfigAutoStore";