+++ packages/imperative/src/logger/__tests__/LoggerUtils.unit.test.ts
@@ -10,7 +10,7 @@
 */
 
 import { EnvironmentalVariableSettings } from "../../imperative/src/env/EnvironmentalVariableSettings";
-import { LoggerUtils } from "../../logger";
+import { LoggerUtils } from "../src/LoggerUtils";
 import { ImperativeConfig } from "../../utilities/src/ImperativeConfig";
 
 describe("LoggerUtils tests", () => {