+++ packages/imperative/src/rest/src/client/doc/IHeaderContent.ts
@@ -9,6 +9,11 @@
 *
 */
 
+/**
+ * Interface to add index signature to http headers
+ * @export
+ * @interface IHeaderContent
+ */
 export interface IHeaderContent {
     [key: string]: string;
 }